move save_and_compress_svg() to pymatviz/utils.py
was in examples/generate_assets.py before
