+++ pymatviz/utils.py
@@ -1,7 +1,9 @@
 from __future__ import annotations
 
 import ast
+import subprocess
 from os.path import abspath, dirname
+from shutil import which
 from typing import Any, Literal, Sequence, Union
 
 import matplotlib.pyplot as plt
@@ -241,3 +243,34 @@ def add_identity_line(
     )
 
     return fig
+
+
+def save_and_compress_svg(filename: str, fig: Figure | None = None) -> None:
+    """Save Plotly figure as SVG and HTML to assets/ folder. Compresses SVG file with
+    svgo CLI if available in PATH.
+
+    Args:
+        fig (Figure): Plotly Figure instance.
+        filename (str): Name of SVG file (w/o extension).
+
+    Raises:
+        ValueError: If fig is None and plt.gcf() is empty.
+    """
+    assert not filename.endswith(".svg"), f"{filename = } should not include .svg"
+    filepath = f"{ROOT}/assets/{filename}.svg"
+
+    if isinstance(fig, Figure):
+        fig.write_image(filepath)
+    elif fig is None:
+        if len(plt.gcf().axes) == 0:
+            raise ValueError(
+                "No figure passed explicitly and plt.gcf() contains no axes. "
+                "Did you forget to pass a plotly figure instance?"
+            )
+        plt.savefig(filepath, bbox_inches="tight")
+        plt.close()
+    else:
+        raise TypeError(f"{fig = } should be a Plotly Figure or Matplotlib Figure")
+
+    if (svgo := which("svgo")) is not None:
+        subprocess.run([svgo, "--multipass", filepath])
